build: migrate to use buildSrc and convention plugins
See:
* [build sources](https://docs.gradle.org/current/userguide/organizing_gradle_projects.html#sec:build_sources)
* [convention plugins](https://docs.gradle.org/current/samples/sample_convention_plugins.html)
