+++ clients/graphql-kotlin-client-jackson/build.gradle.kts
@@ -1,5 +1,9 @@
 description = "GraphQL client serializer based on Jackson"
 
+plugins {
+    id("com.expediagroup.graphql.conventions")
+}
+
 dependencies {
     api(project(path = ":graphql-kotlin-client"))
     api(libs.jackson)
