+++ clients/graphql-kotlin-client-serialization/build.gradle.kts
@@ -1,7 +1,9 @@
 description = "GraphQL client serializer based on kotlinx.serialization"
 
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: remove once KTIJ-19369 / Gradle#22797 is fixed
 plugins {
-    kotlin("plugin.serialization")
+    id("com.expediagroup.graphql.conventions")
+    alias(libs.plugins.kotlin.serialization)
 }
 
 dependencies {
