+++ clients/graphql-kotlin-client/build.gradle.kts
@@ -1,5 +1,9 @@
 description = "A lightweight typesafe GraphQL HTTP Client"
 
+plugins {
+    id("com.expediagroup.graphql.conventions")
+}
+
 dependencies {
     api(libs.kotlinx.coroutines.core)
 }
