+++ examples/client/gradle-client/build.gradle.kts
@@ -4,10 +4,12 @@ import com.expediagroup.graphql.plugin.gradle.graphql
 
 description = "Example usage of Gradle plugin to generate GraphQL Kotlin Client"
 
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: remove once KTIJ-19369 / Gradle#22797 is fixed
 plugins {
-    application
-    kotlin("plugin.serialization")
+    id("com.expediagroup.graphql.examples.conventions")
     id("com.expediagroup.graphql")
+    application
+    alias(libs.plugins.kotlin.serialization)
 }
 
 val ktorVersion: String by project
