+++ examples/client/server/build.gradle.kts
@@ -1,3 +1,10 @@
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: remove once KTIJ-19369 / Gradle#22797 is fixed
+plugins {
+    id("com.expediagroup.graphql.examples.conventions")
+    alias(libs.plugins.kotlin.spring)
+    alias(libs.plugins.spring.boot)
+}
+
 plugins {
     kotlin("plugin.spring")
     id("org.springframework.boot")
