+++ examples/gradle.properties
@@ -1,4 +1,4 @@
-group = com.expediagroup
+group = com.expediagroup.examples
 version = 7.0.0-SNAPSHOT
 
 # build config
