+++ plugins/schema/graphql-kotlin-hooks-provider/build.gradle.kts
@@ -1,5 +1,9 @@
 description = "Module containing SchemaGeneratorHooksProvider Service Provider Interface (SPI)"
 
+plugins {
+    id("com.expediagroup.graphql.conventions")
+}
+
 dependencies {
     implementation(project(":graphql-kotlin-schema-generator"))
 }
