+++ servers/graphql-kotlin-server/build.gradle.kts
@@ -2,8 +2,10 @@ import kotlinx.benchmark.gradle.JvmBenchmarkTarget
 
 description = "Common code for running a GraphQL server in any HTTP server framework"
 
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: remove once KTIJ-19369 / Gradle#22797 is fixed
 plugins {
-    id("org.jetbrains.kotlinx.benchmark")
+    id("com.expediagroup.graphql.conventions")
+    alias(libs.plugins.benchmark)
 }
 
 dependencies {
