+++ servers/graphql-kotlin-spring-server/build.gradle.kts
@@ -1,8 +1,10 @@
 description = "Spring Boot autoconfiguration library for creating reactive GraphQL server"
 
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: remove once KTIJ-19369 / Gradle#22797 is fixed
 plugins {
-    kotlin("plugin.spring")
-    kotlin("kapt")
+    id("com.expediagroup.graphql.conventions")
+    alias(libs.plugins.kotlin.spring)
+    id(libs.plugins.kotlin.kapt.get().pluginId)
 }
 
 dependencies {
