+++ settings.gradle.kts
@@ -1,30 +1,3 @@
-pluginManagement {
-    val detektVersion: String by settings
-    val dokkaVersion: String by settings
-    val kotlinVersion: String by settings
-    val kotlinxBenchmarkVersion: String by settings
-    val ktlintPluginVersion: String by settings
-    val mavenPluginDevelopmentVersion: String by settings
-    val nexusPublishPluginVersion: String by settings
-    val pluginPublishPluginVersion: String by settings
-    val springBootVersion: String by settings
-
-    plugins {
-        kotlin("jvm") version kotlinVersion
-        kotlin("kapt") version kotlinVersion
-        kotlin("plugin.serialization") version kotlinVersion
-        kotlin("plugin.spring") version kotlinVersion
-        id("com.gradle.plugin-publish") version pluginPublishPluginVersion
-        id("de.benediktritter.maven-plugin-development") version mavenPluginDevelopmentVersion
-        id("io.github.gradle-nexus.publish-plugin") version nexusPublishPluginVersion
-        id("io.gitlab.arturbosch.detekt") version detektVersion
-        id("org.jetbrains.dokka") version dokkaVersion
-        id("org.jetbrains.kotlinx.benchmark") version kotlinxBenchmarkVersion
-        id("org.jlleitschuh.gradle.ktlint") version ktlintPluginVersion
-        id("org.springframework.boot") version springBootVersion
-    }
-}
-
 rootProject.name = "graphql-kotlin"
 
 // Clients
